finished with channels routes

+++ backend/monolithic/internal/services/channels_service.go
@@ -3,6 +3,7 @@ package services
 import (
 	"MyChatApp/monolithic/internal/models"
 	"MyChatApp/monolithic/internal/repositories/db"
+	"errors"
 	"time"
 
 	"github.com/gocql/gocql"
@@ -56,15 +57,15 @@ func (c *ChannelService) GetUsersByChannel(chn_id string) ([]models.User, error)
 	return c.usersRepo.FindByIDs(userIDs)
 }
 
-func (c *ChannelService) CreateChannel(creator_id, name, _type string) error {
+func (c *ChannelService) CreateChannel(creator_id, name, _type string) (*models.Channel, error) {
 	cid, err := gocql.RandomUUID()
 	if err != nil {
-		return err
+		return nil, err
 	}
 
 	creatorID, err := gocql.ParseUUID(creator_id)
 	if err != nil {
-		return err
+		return nil, err
 	}
 
 	var chn = models.Channel{
@@ -75,24 +76,29 @@ func (c *ChannelService) CreateChannel(creator_id, name, _type string) error {
 		CreatedAt: time.Now(),
 	}
 
-	return c.channelsRepo.CreateChannel(&chn)
+	err = c.channelsRepo.CreateChannel(&chn)
+	if err != nil {
+		return nil, err
+	}
+
+	return &chn, nil
 }
 
 // TODO: maybe make role a parameter?
-func (c *ChannelService) CreateChannelMembership(user_id, channel_id string) error {
+func (c *ChannelService) CreateChannelMembership(user_id, channel_id string) (*models.ChannelMembership, error) {
 	uid, err := gocql.ParseUUID(user_id)
 	if err != nil {
-		return err
+		return nil, err
 	}
 
 	cid, err := gocql.ParseUUID(channel_id)
 	if err != nil {
-		return err
+		return nil, err
 	}
 
 	chn, err := c.channelsRepo.GetChannelByID(cid)
 	if err != nil {
-		return err
+		return nil, err
 	}
 
 	chnMem := models.ChannelMembership{
@@ -102,7 +108,12 @@ func (c *ChannelService) CreateChannelMembership(user_id, channel_id string) err
 		JoinedAt:    time.Now(),
 	}
 
-	return c.channelsRepo.CreateChannelMembership(uid, &chnMem, "member")
+	err = c.channelsRepo.CreateChannelMembership(uid, &chnMem, "member")
+	if err != nil {
+		return nil, err
+	}
+
+	return &chnMem, nil
 }
 
 func (c *ChannelService) ModifyChannelMembership(user_id, channel_id, role string) error {
@@ -131,15 +142,25 @@ func (c *ChannelService) ModifyChannelMembership(user_id, channel_id, role strin
 	return c.channelsRepo.ModifyChannelMembership(uid, &chnMem, role)
 }
 
-func (c *ChannelService) DeleteChannel(channel_id string) error {
+func (c *ChannelService) DeleteChannel(user_id, channel_id string) error {
 	cid, err := gocql.ParseUUID(channel_id)
 	if err != nil {
 		return err
 	}
 
+	role, err := c.channelsRepo.GetMemberRole(channel_id, user_id)
+	if err != nil {
+		return err
+	}
+
+	if role != "owner" {
+		return errors.New("Only owners can delete channels")
+	}
+
 	return c.channelsRepo.DeleteChannel(cid)
 }
 
+// TODO: If 1 person left in DM, delete DM channel too? if 0 people left in channel, delete channel?
 func (c *ChannelService) DeleteChannelMembership(user_id, channel_id string) error {
 	uid, err := gocql.ParseUUID(user_id)
 	if err != nil {
